[mm] userfaultfd: propagate the full address in THP faults
Message-id: <1437174639-28447-38-git-send-email-aarcange@redhat.com>
Patchwork-id: 122116
O-Subject: [RHEL7.2 PATCH 37/39] userfaultfd: propagate the full address in THP faults
Bugzilla: 965657
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

The THP faults were not propagating the original fault address. The latest
version of the API with uffd.arg.pagefault.address is supposed to propagate the
full address through THP faults.

This was not a kernel crashing bug and it wouldn't risk to corrupt
user memory, but it would cause a SIGBUS failure because the wrong page was
being copied.

For various reasons this wasn't easily reproducible in the qemu
workload, but the strestest exposed the problem immediately.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>