[fs] ovl: don't remove non-empty opaque directory
Message-id: <30189.1437401773@warthog.procyon.org.uk>
Patchwork-id: 122177
O-Subject: [RHEL7 PATCH 25a/28] ovl: don't remove non-empty opaque directory
Bugzilla: 1178942
RH-Acked-by: Steve Best <sbest@redhat.com>

From: Miklos Szeredi <mszeredi@suse.cz>

When removing an opaque directory we can't just call rmdir() to check for
emptiness, because the directory will need to be replaced with a whiteout.
The replacement is done with RENAME_EXCHANGE, which doesn't check
emptiness.

Solution is just to check emptiness by reading the directory.  In the
future we could add a new rename flag to check for emptiness even for
RENAME_EXCHANGE to optimize this case.

Reported-by: Vincent Batts <vbatts@gmail.com>
Signed-off-by: Miklos Szeredi <mszeredi@suse.cz>
Tested-by: Jordi Pujol Palomer <jordipujolp@gmail.com>
Fixes: 263b4a0fee43 ("ovl: dont replace opaque dir")
Cc: <stable@vger.kernel.org> # v4.0+
Original-commit: d377c5eb54dd05aeb3094b7740252d19ba7791f7
RH-Bugzilla: 1178942
Signed-off-by: Rafael Aquini <aquini@redhat.com>