[x86] perf: Fix CQM feature detection
Message-id: <20150722165325.GC22824@krava.brq.redhat.com>
Patchwork-id: 122329
O-Subject: [PATCHv2 RHEL7.2 BZ1036948] perf x86: Fix CQM feature detection
Bugzilla: 1036948
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Oleg Nesterov <oleg@redhat.com>
RH-Acked-by: Jerome Marchand <jmarchan@redhat.com>

Bugzilla: 1036948
https://bugzilla.redhat.com/show_bug.cgi?id=1036948

upstream
========
N/A - RHEL specific change

brew
====
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=9565341

description
===========
Adding initialization of rh_boot_cpu_data with
detected data, because it's used later on in
CQM init code. It's leftover from:

  33748b762cd0 [x86] perf: Fix bug in unused code

The way the code currently works is that get_cpu_cap accidentally writes to
rh_cpu_data(0) before it is committed.  This works because it was
accidentally forgotten to write the rh_boot_cpu_data to rh_cpu_data(0) in
smp_store_boot_cpu_info (which would have overwritten the real data in
rh_cpu_data(0) with junk).

However, because the cqm code actually relies on rh_boot_cpu_data as a
reference when initializing itself on other cpus, the junk data did not line
up correctly with the read data exhibiting a failure on init, I think.

So the fix is to use the correct struct in get_cpu_cap (this is hackish because
of the kabi guarantee).  This fills in the rh_boot_cpu_data for use by
other folks.  Meanwhile, you also need to make sure rh_boot_cpu_data is
written correctly to rh_cpu_data(0), which is what the second half of the
fix does.

Signed-off-by: Rafael Aquini <aquini@redhat.com>