[iser-target] Remove un-needed rdma_listen backlog
Message-id: <1437767140-16484-19-git-send-email-agrover@redhat.com>
Patchwork-id: 122390
O-Subject: [RHEL7.2 PATCHv3 e-stor 18/23] iser-target: Remove un-needed rdma_listen backlog
Bugzilla: 1136558 1185396
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Rob Evers <revers@redhat.com>

From: Sagi Grimberg <sagig@mellanox.com>

iser target can handle as many connect request as
the fabric sends to it. This backlog should not set as
a back-pressure mechanism (which is not very useful).

isert does need a back-pressure mechanism, but it should
be added in isert by monitoring the number of pending
established connections (will be added in a later stage).

Signed-off-by: Sagi Grimberg <sagig@mellanox.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>