[scsi] qla2xxx: Do not reset adapter if SRB handle is in range
Message-id: <1438007050-3347-3-git-send-email-cdupuis@redhat.com>
Patchwork-id: 122538
O-Subject: [PATCH RHEL 7.2 e-stor 2/4] qla2xxx: Do not reset adapter if SRB handle is in range.
Bugzilla: 1188168
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Chad Dupuis <chad.dupuis@qlogic.com>

Upstream posting pending

If an SRB is NULL but the handle is in range just drop the
command instead of also resetting the adapter. If the handle
is in range then the command was valid at some point and may
have been aborted. Resetting the adapter can lead to extended
recovery times in this case.

Signed-off-by: Chad Dupuis <chad.dupuis@qlogic.com>
Signed-off-by: Himanshu Madhani <himanshu.madhani@qlogic.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>