[fs] nfsd4: disallow ALLOCATE with special stateids
Message-id: <1438285830-10681-16-git-send-email-bfields@redhat.com>
Patchwork-id: 122765
O-Subject: [Patch RHEL7 fs 15/17] nfsd4: disallow ALLOCATE with special stateids
Bugzilla: 1079385
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Andreas Grünbacher <agruenba@redhat.com>

vfs_fallocate will hit a NULL dereference if the client tries an
ALLOCATE or DEALLOCATE with a special stateid.  Fix that.  (We also
depend on the open to have broken any conflicting leases or delegations
for us.)

(If it turns out we need to allow special stateid's then we could do a
temporary open here in the special-stateid case, as we do for read and
write.  For now I'm assuming it's not necessary.)

Fixes: 95d871f03cae "nfsd: Add ALLOCATE support"
Cc: stable@vger.kernel.org
Cc: Anna Schumaker <Anna.Schumaker@Netapp.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream: 5ba4a25ab7b13be528b23f85182f4d09cf7f71ad
Signed-off-by: Rafael Aquini <aquini@redhat.com>