[x86] kprobes: Fix conflict between jprobes and function graph tracing
Message-id: <1438344120-30528-1-git-send-email-jolsa@redhat.com>
Patchwork-id: 122871
O-Subject: [PATCH RHEL7.2 BZ1242825] ftrace/jprobes/x86: Fix conflict between jprobes and function graph tracing
Bugzilla: 1242825
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Oleg Nesterov <oleg@redhat.com>
RH-Acked-by: Jerome Marchand <jmarchan@redhat.com>

From: "Steven Rostedt (Red Hat)" <rostedt@goodmis.org>

Bugzilla: 1242825
https://bugzilla.redhat.com/show_bug.cgi?id=1242825

upstream
========
commit 237d28db036e411f22c03cfd5b0f6dc2aa9bf3bc
Author: Steven Rostedt (Red Hat) <rostedt@goodmis.org>
Date:   Mon Jan 12 12:12:03 2015 -0500

brew
====
https://brewweb.devel.redhat.com/taskinfo?taskID=9618091

tested
======
As specified in the BZ and changelog.

description
===========
If the function graph tracer traces a jprobe callback, the system will
crash. This can easily be demonstrated by compiling the jprobe
sample module that is in the kernel tree, loading it and running the
function graph tracer.

 # modprobe jprobe_example.ko
 # echo function_graph > /sys/kernel/debug/tracing/current_tracer
 # ls

The first two commands end up in a nice crash after the first fork.
(do_fork has a jprobe attached to it, so "ls" just triggers that fork)

The problem is caused by the jprobe_return() that all jprobe callbacks
must end with. The way jprobes works is that the function a jprobe
is attached to has a breakpoint placed at the start of it (or it uses
ftrace if fentry is supported). The breakpoint handler (or ftrace callback)
will copy the stack frame and change the ip address to return to the
jprobe handler instead of the function. The jprobe handler must end
with jprobe_return() which swaps the stack and does an int3 (breakpoint).
This breakpoint handler will then put back the saved stack frame,
simulate the instruction at the beginning of the function it added
a breakpoint to, and then continue on.

For function tracing to work, it hijakes the return address from the
stack frame, and replaces it with a hook function that will trace
the end of the call. This hook function will restore the return
address of the function call.

If the function tracer traces the jprobe handler, the hook function
for that handler will not be called, and its saved return address
will be used for the next function. This will result in a kernel crash.

To solve this, pause function tracing before the jprobe handler is called
and unpause it before it returns back to the function it probed.

Some other updates:

Used a variable "saved_sp" to hold kcb->jprobe_saved_sp. This makes the
code look a bit cleaner and easier to understand (various tries to fix
this bug required this change).

Note, if fentry is being used, jprobes will change the ip address before
the function graph tracer runs and it will not be able to trace the
function that the jprobe is probing.

Signed-off-by: Rafael Aquini <aquini@redhat.com>