[fs] inotify: don't add consecutive overflow events to the queue
Message-id: <55C0EAC1.70109@redhat.com>
Patchwork-id: 123069
O-Subject: [RHEL7.2 PATCH] [fs] inotify: don't add consecutive overflow events to the queue
Bugzilla: 1168465
RH-Acked-by: Eric Paris <eparis@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1168465
Upstream Status: N/A - upstream has diverged significantly
Build Info: Local
Tested: Locally, with reproducer in bug

This is a RHEL7-specific fix for inotify, essentially a DOS
where if userspace does a tight loop of

	for (;;) {
		wd = inotify_add_watch(notify_fd, ... );
		ret = inotify_rm_watch(notify_fd, wd);
	}

the fsnotify_event_holder cache will grow unbounded, eventually
OOMing the machine.

This is because after we have exceeded the maximum queue of
events, we start using the q_overflow_event instead, and the
in-event holder is quickly consumed, so we start allocating
holders for each overflow event that we continue to tack onto
the end of the queue.

Fix this by not adding consecutive overflow events to the end of
the queue; if one is already there, just take a ref on it and
return it.

This is RHEL7-specific; commit 7053aee26 reworked things
significantly, and the cache in question doesn't even exist
anymore.

Thanks to eparis for discussing this with me, suggesting this
approach, filling me in on inotify background, and pre-reviewing
this patch.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>