[vfio] Fix lockdep issue
Message-id: <20150804193428.5818.25246.stgit@gimli.home>
Patchwork-id: 123086
O-Subject: [RHEL7.2 kernel PATCH] vfio: Fix lockdep issue
Bugzilla: 1250237
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1250237
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=9638761
Upstream: 4bc94d5dc95d3a2159d821b39277477e69507f67

When we open a device file descriptor, we currently have the
following:

vfio_group_get_device_fd()
  mutex_lock(&group->device_lock);
    open()
    ...
    if (ret)
      release()

If we hit that error case, we call the backend driver release path,
which for vfio-pci looks like this:

vfio_pci_release()
  vfio_pci_disable()
    vfio_pci_try_bus_reset()
      vfio_pci_get_devs()
        vfio_device_get_from_dev()
          vfio_group_get_device()
            mutex_lock(&group->device_lock);

Whoops, we've stumbled back onto group.device_lock and created a
deadlock.  There's a low likelihood of ever seeing this play out, but
obviously it needs to be fixed.  To do that we can use a reference to
the vfio_device for vfio_group_get_device_fd() rather than holding the
lock.  There was a loop in this function, theoretically allowing
multiple devices with the same name, but in practice we don't expect
such a thing to happen and the code is already aborting from the loop
with break on any sort of error rather than continuing and only
parsing the first match anyway, so the loop was effectively unused
already.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Fixes: 20f300175a1e ("vfio/pci: Fix racy vfio_device_get_from_dev() call")
Reported-by: Joerg Roedel <joro@8bytes.org>
Tested-by: Joerg Roedel <jroedel@suse.de>
---

Testing: Used prlimit to set the open file limit for a VM to fail at
exactly the point that triggers this open()/release() sequence, then
attempted to hot-add a device.  qemu-kvm process hangs and hung task
trace is triggered.  All of this is documented in the bz for QE.  With
fix, a graceful open file limit error occurs, the hot-add fails, and
the VM continues normally.

Signed-off-by: Rafael Aquini <aquini@redhat.com>