[powerpc] kvm: book3s: correct width in XER handling
Message-id: <1439296825-24515-1-git-send-email-dgibson@redhat.com>
Patchwork-id: 123299
O-Subject: [RHEL-7.2 kernel PATCH] KVM: PPC: Book3S: correct width in XER handling
Bugzilla: 1178502
RH-Acked-by: Laurent Vivier <lvivier@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>

In 64 bit kernels, the Fixed Point Exception Register (XER) is a 64
bit field (e.g. in kvm_regs and kvm_vcpu_arch) and in most places it is
accessed as such.

This patch corrects places where it is accessed as a 32 bit field by a
64 bit kernel.  In some cases this is via a 32 bit load or store
instruction which, depending on endianness, will cause either the
lower or upper 32 bits to be missed.  In another case it is cast as a
u32, causing the upper 32 bits to be cleared.

This patch corrects those places by extending the access methods to
64 bits.

Signed-off-by: Sam Bobroff <sam.bobroff@au1.ibm.com>
Reviewed-by: Laurent Vivier <lvivier@redhat.com>
Reviewed-by: Thomas Huth <thuth@redhat.com>
Reviewed-by: David Gibson <david@gibson.dropbear.id.au>
Tested-by: Thomas Huth <thuth@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1178502
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=9667869
Upstream: Not upstream :(

This has been posted upstream, but it's being very slow to merge despite
repeated pings to the maintainer, Alex Graf.

It is, however, a vital fix for register corruption of KVM guests on Power,
so I think we need to look at merging it downstream anyway.

Signed-off-by: David Gibson <dgibson@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>