[unisys] Fix improper use of NETDEV_TX_BUSY
Message-id: <1439308494-5033-24-git-send-email-earfvids@redhat.com>
Patchwork-id: 123321
O-Subject: [RHEL7.2 PATCH BZ 1232018 23/38] staging: unisys: Fix improper use of NETDEV_TX_BUSY
Bugzilla: 1232018
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

From: Neil Horman <nhorman@redhat.com>

Using NETDEV_TX_BUSY is tricky.  Its meant for situations where the error
in question is transient and quickly resolved.  But the driver rarely is
able to know that to a certainty.  And in the case of visornic, it just
uses it without any care for that, in the hopes that it won't loose frames,
even if the problem is that the skb is somehow malformed for the hardware.
If we get one of those kinds of skbs, NETDEV_TX_BUSY will just cause us to
spin, processing the same error over and over.

Fix it by dropping the frame, stopping the queue where appropriate, and
returning NETDEV_TX_OK

Signed-off-by: Neil Horman <nhorman@redhat.com>
Signed-off-by: Benjamin Romer <benjamin.romer@unisys.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit f6346ad662e44d7086bc3c18fc19a33f0b95b576)

Cc: Erik Arfvidson <earfvids@redhat.com>
Cc: Neil Horman <nhorman@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>