[scsi] hpsa: add masked physical devices into h->dev array
Message-id: <1439394090-11832-4-git-send-email-jszczype@redhat.com>
Patchwork-id: 123461
O-Subject: [RHEL7.2 e-stor PATCH 03/46] hpsa: add masked physical devices into h->dev[] array
Bugzilla: 1227171
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Rob Evers <revers@redhat.com>

From: Stephen Cameron <stephenmcameron@gmail.com>

Cache the ioaccel handle so that when we need to abort commands sent
down the ioaccel2 path, we can look up the LUN ID in h->dev[] instead of
having to do I/O to the controller.

Add a field to elements in h->dev[] to keep track of how the device is exposed
to the SCSI mid layer: Not at all, without an upper level driver
(no_uld_attach) or normally exposed.

Since masked physical devices are now present in h->dev[] array
it would be perfectly possible to do

	echo scsi add-single-device 2 2 0 0 > /proc/scsi/scsi

and bring them online.  This was previously not allowed for masked
physical devices.

Ensure that the mapping of physical disks to logical drives gets updated in a
consistent way when a RAID migration occurs and is not touched until updates
to it are complete.

now instead of doing CISS_REPORT_PHYSICAL to get the LUNID for
the physical disk in hpsa_get_pdisk_of_ioaccel2(), just get
it out of h->dev[] where we already have it cached.

do not touch phys_disk[] for ioaccel enabled logical drives during rescan

Reviewed-by: Scott Teel <scott.teel@pmcs.com>
Reviewed-by: Kevin Barnett <kevin.barnett@pmcs.com>
Reviewed-by: Tomas Henzl <thenzl@redhat.com>
Reviewed-by: Hannes Reinecke <hare@Suse.de>
Signed-off-by: Don Brace <don.brace@pmcs.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: James Bottomley <JBottomley@Odin.com>
(cherry picked from commit 41ce4c355765a5e0a8e1c8ff8d7257160bbae93d)
Signed-off-by: Rafael Aquini <aquini@redhat.com>

Conflicts:
	drivers/scsi/hpsa.c

Conflict occurred with hpsa_slave_alloc().  The lines being removed in that
function by the upstream commit do not match the lines that are there in
the driver for RHEL7.2.   Difference are there due to RHEL7 not having
scsi_change_queue_depth().   Modification was manually made to match what is
upstream, and to continue to use scsi_adjust_queue_depth in the
hpsa_slave_configure function.