[netdrv] ixgbevf: add support for reporting RSS key and hash table for X550
Message-id: <1439583501-12474-35-git-send-email-jogreene@redhat.com>
Patchwork-id: 124080
O-Subject: [RHEL7.2 bz1173787 v2 34/35] ixgbevf: add support for reporting RSS key and hash table for X550
Bugzilla: 1173787
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

BZ 1173787
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=9415509
Tested: In beaker. See BZ for details
Backport net-next commit

This patch extends the reporting of the RSS key and hash table by
adding support for X550 VFs. The difference is that X550 VFs have
their own registers for RSS key and indirection table, so there is
no need to query the PF.

The RSS key and indirection table are stored in the adapter structure
during the configuration of VFRSSRK and VFRETA which in turn can be
used in ethtool for reporting.

The logic for writing VFRETA is also changed to make sure that the
indirection table is reported correctly.

In addition this patch adds defines for the VFRETA entries and number
of VFRSSRK registers as well as some whitespace cleanups.

Reported-by: Vlad Zolotarov <vladz@cloudius-systems.com>
Signed-off-by: Emil Tantilov <emil.s.tantilov@intel.com>
Tested-by: Phil Schmitt <phillip.j.schmitt@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit 9cba434f630a972b47327ae3d014445033166206)
Signed-off-by: Rafael Aquini <aquini@redhat.com>