[infiniband] revert "ib/hfi1: Switch to topology_sibling_mask()"
Message-id: <1440002030-2889-54-git-send-email-dsowa@redhat.com>
Patchwork-id: 124166
O-Subject: [RHEL 7.2 PATCH bz1173291 53/57] Revert "IB/hfi1: Switch to topology_sibling_mask()"
Bugzilla: 1173291
RH-Acked-by: Doug Ledford <dledford@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: David Sowa <david.sowa@intel.com>

Upstream driver was written in a private git repo.  When submitted
upstream, it was submitted as files, not as all of the individual
commits used to write the driver.  During writing of the driver, updates
were made to make the driver use modern kernel infrastructure that is
not present in RHEL.  Consequently, it was necessary to revert some of
the internal git commits from when the driver was being written in order
to back out updates to newer infrastructure that RHEL does not have.
Those commits can not be found upstream because of how the driver was
submitted and can be considered RHEL specific patches.

Signed-off-by: David Sowa <dsowa@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>