[netdrv] ixgbe: Cleanup FDB handling code
Message-id: <1440100662-9783-56-git-send-email-jogreene@redhat.com>
Patchwork-id: 124306
O-Subject: [RHEL7.2 bz1173786 v3 055/148] ixgbe: Cleanup FDB handling code
Bugzilla: 1173786
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

Bugzilla: 1173786
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=9415509
Tested: In beaker. See BZ for details

This change makes it so that the behavior for FDB handling is consistent
between both the SR-IOV and non-SR-IOV cases.  The main change here is that we
perform bounds checking on the number of SR-IOV addresses regardless of if
SR-IOV is enabled or not as we can only support a certain number of addresses
in the hardware.

Signed-off-by: Alexander Duyck <alexander.h.duyck@intel.com>
Tested-by: Phil Schmitt <phillip.j.schmitt@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit bcfd3432d1625ef749a8697d194010d0c75b97c9)
Signed-off-by: Rafael Aquini <aquini@redhat.com>