[netdrv] ixgbe: Refactor busy poll socket code to address multiple issues
Message-id: <1440100662-9783-62-git-send-email-jogreene@redhat.com>
Patchwork-id: 124315
O-Subject: [RHEL7.2 bz1173786 v3 061/148] ixgbe: Refactor busy poll socket code to address multiple issues
Bugzilla: 1173786
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

Bugzilla: 1173786
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=9415509
Tested: In beaker. See BZ for details

This change addresses several issues in the current ixgbe implementation of
busy poll sockets.

First was the fact that it was possible for frames to be delivered out of
order if they were held in GRO.  This is addressed by flushing the GRO buffers
before releasing the q_vector back to the idle state.

The other issue was the fact that we were having to take a spinlock on
changing the state to and from idle.  To resolve this I have replaced the
state value with an atomic and use atomic_cmpxchg to change the value from
idle, and a simple atomic set to restore it back to idle after we have
acquired it.  This allows us to only use a locked operation on acquiring the
vector without a need for a locked operation to release it.

Signed-off-by: Alexander Duyck <alexander.h.duyck@intel.com>
Tested-by: Phil Schmitt <phillip.j.schmitt@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit adc810900a703ee78fe88fd65e086d359fec04b2)
Signed-off-by: Rafael Aquini <aquini@redhat.com>