[fs] locks: plumb a "priv" pointer into the setlease routines
Message-id: <1440612355-19877-1-git-send-email-steved@redhat.com>
Patchwork-id: 124491
O-Subject: [RHEL7.2 fs PATCH 05/26 v3] locks: plumb a "priv" pointer into the setlease routines
Bugzilla: 1195496
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Andreas Grünbacher <agruenba@redhat.com>

From: Jeff Layton <jlayton@primarydata.com>

In later patches, we're going to add a new lock_manager_operation to
finish setting up the lease while still holding the i_lock.  To do
this, we'll need to pass a little bit of info in the fcntl setlease
case (primarily an fasync structure). Plumb the extra pointer into
there in advance of that.

We declare this pointer as a void ** to make it clear that this is
private info, and that the caller isn't required to set this unless
the lm_setup specifically requires it.

Signed-off-by: Jeff Layton <jlayton@primarydata.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Upstream Id: e6f5c78930e409f3a6b37f5484313a416359ac7f

Signed-off-by: Rafael Aquini <aquini@redhat.com>