[pci] Add ACS quirks for Intel I219-LM/V
Message-id: <20150827142236.20042.8769.stgit@gimli.home>
Patchwork-id: 124499
O-Subject: [RHEL7.2 kernel PATCH] PCI: Add ACS quirks for Intel I219-LM/V
Bugzilla: 1252128
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Myron Stowe <mstowe@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>

Bugzilla: 1252128
Brew: https://brewweb.devel.redhat.com/taskinfo?taskID=9752279
Upstream: 95e16587ccd0f330e349360de048ebd4432d4b2c (linux-next)

The Intel 100-series chipset now includes the integrated Ethernet as part
of a multifunction package.  The Ethernet function does not include native
ACS support, but Intel confirms that the device is not capable of peer-to-
peer within the package.  We can therefore quirk it to expose the
isolation.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
Acked-by: John Ronciak <john.ronciak@gmail.com>
---

Quite possibly too late for this, but it's a trivial addition to an
existing quirk, so I'll post and let pm decided whether we want it.
This makes the integrated NIC on skylake chipsets assignable to a VM.

Signed-off-by: Rafael Aquini <aquini@redhat.com>