[fs] ext4: fix potential use after free in __ext4_journal_stop
Message-id: <1440774866-2967-2-git-send-email-lczerner@redhat.com>
Patchwork-id: 124514
O-Subject: [RHEL 7.2 fs PATCH v2 2/2] ext4: fix potential use after free in __ext4_journal_stop
Bugzilla: 1210588
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>
RH-Acked-by: Brian Foster <bfoster@redhat.com>

This commit does not have it's upstream counterpart yet. The reason is
that we've discover a bug in the previous commit "ext4: fix NULL pointer
dereference when journal restart fails" however due to a tight schedule
we can't afford to wait for upstream inclusion.

The same patch will be sent and hopefully accepted upstream later.

There is a use-after-free possibility in __ext4_journal_stop() in the
case that we free the handle in the first jbd2_journal_stop() because
we're referencing handle->h_err afterwards. This is wrong. Fix it by
storing the handle->h_err value beforehand and avoid referencing
potentially freed handle.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>