[x86] mce: Do not panic when single core has reached a timeout
Message-id: <1441201898-21681-1-git-send-email-prarit@redhat.com>
Patchwork-id: 124677
O-Subject: [RHEL7.2 PATCH BZ 1195534] mce: Do not panic when single core has reached a timeout
Bugzilla: 1195534
RH-Acked-by: Seth Jennings <sjenning@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Dave Young <dyoung@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1195534

Brew: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=9737348

RHEL_only.

The discussion around this issue is ongoing upstream.  The problem is that
on some Intel boxes the kdump kernel will panic here due to an mce
timeout.  Unfortunately this is being seen on systems that are already
supported by Red Hat and we cannot risk this type of regression.

Dave Young has requested that I back out this change (which is a trivial
change) and I've agreed to do it for 7.2 so that we can make sure that
kdump continues to work correctly.  I will backport a more appropriate fix
for 7.3.

Backing out this change doesn't seem to hurt anything and keeps the code
in line with what was released in 7.1 so there is no functionality loss
in RHEL7.

Successfully tested by NEC,
https://bugzilla.redhat.com/show_bug.cgi?id=1195534#c32

Cc: Dave Young <dyoung@redhat.com>
Cc: Seth Jennings <sjenning@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>