[fs] cifs: Terminate cifs_smb_version_tokens with NULL
Message-id: <1441279386-29469-1-git-send-email-sprabhu@redhat.com>
Patchwork-id: 124691
O-Subject: [RHEL7 PATCH fs] [bz1247446]cifs: Terminate cifs_smb_version_tokens with NULL
Bugzilla: 1247446
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>

Fixes bz 1247446.

The table with the set of allowed options tokens must be terminated with
a token whose pattern is set to the NULL pointer.

This wasn't done earlier. Passing an illegal version string when
attempting to mount a share on ppc64 machines causes it to Oops.

This is a partial backport of upstream commit
5f7fbf733c9dd6b58d40fa394313a5949b696f84
which adds support for version 3.1.1.

A test kernel built with this patch was successfully tested with a ppc64
machine.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Reported-by: Xiaoli Feng <xifeng@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>