[fs] Set UID in sess_auth_rawntlmssp_authenticate too
Message-id: <1441372352-23633-1-git-send-email-sprabhu@redhat.com>
Patchwork-id: 124718
O-Subject: [RHEL7 PATCH fs] [bz1258490]Set UID in sess_auth_rawntlmssp_authenticate too
Bugzilla: 1258490
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

Fixes bz 1258490.

We need to backport the fix for an regression which was introduced
upstream when cleaning up the session setup code. I backported the code
cleanups to RHEL 7 but didn't backport this fix which came in later in
the upstream tree.

Original summary follows
--
A user complained that they were unable to login to their cifs share
after a kernel update. From the wiretrace we can see that the server
returns different UIDs as response to NTLMSSP_NEGOTIATE and NTLMSSP_AUTH
phases.

With changes in the authentication code, we no longer set the
cifs_sess->Suid returned in response to the NTLM_AUTH phase and continue
to use the UID sent in response to the NTLMSSP_NEGOTIATE phase. This
results in the server denying access to the user when the user attempts
to do a tcon connect.

See https://bugzilla.redhat.com/show_bug.cgi?id=1163927

A test kernel containing patch was tested successfully by the user.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Steve French <steve.french@primarydata.com>
(cherry picked from commit ee9bbf465d8bb82ff2081834cb53a0ea7574c423)

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>