[powerpc] kvm: book3s-hv: Exit on H_DOORBELL if HOST_IPI is set
Message-id: <1441687134-2850-3-git-send-email-dgibson@redhat.com>
Patchwork-id: 124733
O-Subject: [RHEL7.2 kernel PATCH 2/2] KVM: PPC: Book3S HV: Exit on H_DOORBELL if HOST_IPI is set
Bugzilla: 1259571
RH-Acked-by: Laurent Vivier <lvivier@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

The code that handles the case when we receive a H_DOORBELL interrupt
has a comment which says "Hypervisor doorbell - exit only if host IPI
flag set".  However, the current code does not actually check if the
host IPI flag is set.  This is due to a comparison instruction that
got missed.

As a result, the current code performs the exit to host only
if some sibling thread or a sibling sub-core is exiting to the
host.  This implies that, an IPI sent to a sibling core in
(subcores-per-core != 1) mode will be missed by the host unless the
sibling core is on the exit path to the host.

This patch adds the missing comparison operation which will ensure
that when HOST_IPI flag is set, we unconditionally exit to the host.

Fixes: 66feed61cdf6
Cc: stable@vger.kernel.org # v4.1+
Signed-off-by: Gautham R. Shenoy <ego@linux.vnet.ibm.com>
Reviewed-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Paul Mackerras <paulus@samba.org>
(cherry picked from commit 06554d9f6cc8f0b5ec903db19726a15dfc7b09d6)

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1259571

Signed-off-by: David Gibson <dgibson@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>