[sound] alsa: hda - Re-add the lost fake mute support
Message-id: <144170153574-9-patchset-jkysela@redhat.com>
Patchwork-id: 124742
O-Subject: [RHEL 7.2 PATCH 08/13] ALSA: hda - Re-add the lost fake mute support
Bugzilla: 1256578
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>

Yet another regression by the transition to regmap cache; for better
usability, we had the fake mute control using the zero amp value for
Conexant codecs, and this was forgotten in the new hda core code.

Since the bits 4-7 are unused for the amp registers (as we follow the
syntax of AMP_GET verb), the bit 4 is now used to indicate the fake
mute.  For setting this flag, snd_hda_codec_amp_update() becomes a
function from a simple macro.  The bonus is that it gained a proper
function description.

Signed-off-by: Takashi Iwai <tiwai@suse.de>

Upstream: a686ec4c5f28eb1b384e4b87b08543155c970072

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>