[md] raid0: update queue parameter in a safer location
Message-id: <1442419029-20109-2-git-send-email-Jes.Sorensen@redhat.com>
Patchwork-id: 124897
O-Subject: [RHEL7.2 PATCH 1/1] md/raid0: update queue parameter in a safer location.
Bugzilla: 1247854
RH-Acked-by: Doug Ledford <dledford@redhat.com>
RH-Acked-by: Xiao Ni <xni@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: NeilBrown <neilb@suse.com>

When a (e.g.) RAID5 array is reshaped to RAID0, the updating
of queue parameters (e.g. max number of sectors per bio) is
done in the wrong place.
It should be part of ->run, but it is actually part of ->takeover.
This means it happens before level_store() calls:

	blk_set_stacking_limits(&mddev->queue->limits);

and so it ineffective.  This can lead to errors from underlying
devices.

So move all the relevant settings out of create_stripe_zones()
and into raid0_run().

As this can lead to a bug-on it is suitable for any -stable
kernel which supports reshape to RAID0.  So 2.6.35 or later.
As the bug has been present for five years there is no urgency,
so no need to rush into -stable.

Fixes: 9af204cf720c ("md: Add support for Raid5->Raid0 and Raid10->Raid0 takeover")
Cc: stable@vger.kernel.org (v2.6.35+ - please delay until 2 weeks after -final release).
Reported-by: Yi Zhang <yizhan@redhat.com>
Signed-off-by: NeilBrown <neilb@suse.com>
(cherry picked from commit 927d881980b74fa653e3992fd4a7283b0e11952b)
Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>

Conflicts:
	drivers/md/raid0.c

Trivial conflict caused by RHEL-7-only patch to default DISCARD
disabled on raid0.