[fs] nfsv4: don't set SETATTR for O_RDONLY|O_EXCL
Message-id: <29663fb4dd9931e43cde5b45dcf49e0499daf168.1443617005.git.bcodding@redhat.com>
Patchwork-id: 125170
O-Subject: [RHEL7.2 PATCH fs] NFSv4: don't set SETATTR for O_RDONLY|O_EXCL
Bugzilla: 1269974
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Dave Wysochanski <dwysocha@redhat.com>

From: NeilBrown <neilb@suse.com>

It is unusual to combine the open flags O_RDONLY and O_EXCL, but
it appears that libre-office does just that.

[pid  3250] stat("/home/USER/.config", {st_mode=S_IFDIR|0700, st_size=8192, ...}) = 0
[pid  3250] open("/home/USER/.config/libreoffice/4-suse/user/extensions/buildid", O_RDONLY|O_EXCL <unfinished ...>

NFSv4 takes O_EXCL as a sign that a setattr command should be sent,
probably to reset the timestamps.

When it was an O_RDONLY open, the SETATTR command does not
identify any actual attributes to change.
If no delegation was provided to the open, the SETATTR uses the
all-zeros stateid and the request is accepted (at least by the
Linux NFS server - no harm, no foul).

If a read-delegation was provided, this is used in the SETATTR
request, and a Netapp filer will justifiably claim
NFS4ERR_BAD_STATEID, which the Linux client takes as a sign
to retry - indefinitely.

So only treat O_EXCL specially if O_CREAT was also given.

Signed-off-by: NeilBrown <neilb@suse.com>
Cc: stable@vger.kernel.org
Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Upstream Id: efcbc04e16dfa95fef76309f89710dd1d99a5453
Bugzilla: 1246394
Signed-off-by: Benjamin Coddington <bcodding@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>