[drm] dp_mst: fixup handling hotplug on port removal
Message-id: <1443681506-20357-4-git-send-email-airlied@redhat.com>
Patchwork-id: 125190
O-Subject: [PATCH 3/6] [PATCH RHEL7.2] drm/dp/mst: fixup handling hotplug on port removal.
Bugzilla: 1251331
RH-Acked-by: Rob Clark <rclark@redhat.com>
RH-Acked-by: Ben Skeggs <bskeggs@redhat.com>
RH-Acked-by: Oded Gabbay <ogabbay@redhat.com>

bugzilla: 1251331
upstream: queued in drm-fixes.

output ports should always have a connector, unless
in the rare case connector allocation fails in the
driver.

In this case we only need to teardown the pdt,
and free the struct, and there is no need to
send a hotplug msg.

In the case were we add the port to the destroy
list we need to send a hotplug if we destroy
any connectors, so userspace knows to reprobe
stuff.

this patch also handles port->connector allocation
failing which should be a rare event, but makes
the code consistent.

Reviewed-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Signed-off-by: Dave Airlie <airlied@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>