[drm] nouveau: fbcon: take runpm reference when userspace has an open fd
Message-id: <1443765262-2749-1-git-send-email-bskeggs@redhat.com>
Patchwork-id: 125278
O-Subject: [RHEL7.2 PATCH] drm/nouveau/fbcon: take runpm reference when userspace has an open fd
Bugzilla: 1176163
RH-Acked-by: Dave Airlie <airlied@redhat.com>
RH-Acked-by: Rob Clark <rclark@redhat.com>

rhbz#1176163

Upstream: queued in nouveau tree, will be sent to 4.3-fixes.

We need to do this in order to prevent accesses to the device while it's
powered down.  Userspace may have an mmap of the fb, and there's no good
way (that I know of) to prevent it from touching the device otherwise.

This fixes some nasty races between runpm and plymouth on some systems,
which result in the GPU getting very upset and hanging to boot.

Signed-off-by: Ben Skeggs <bskeggs@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>