[drm] radeon: update no_64bit_msi flag for certain ASICs
Message-id: <1443867168-9676-1-git-send-email-oded.gabbay@redhat.com>
Patchwork-id: 125398
O-Subject: [RHEL 7.2 PATCH BZ 1262429] drm/radeon: update no_64bit_msi flag for certain ASICs
Bugzilla: 1262429
RH-Acked-by: Rob Clark <rclark@redhat.com>
RH-Acked-by: Dave Airlie <airlied@redhat.com>

This patch updates the no_64bit_msi flag for relevant ASICs. It was
omitted as part of the drm rebase, because that flag was not present
in pci_dev struct at the time of the rebase.

Because the no_64bit_msi flag has now been imported to that structure, we
can update this flag in the radeon driver.

It was found as part of bz #1262429

Signed-off-by: Oded Gabbay <oded.gabbay@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>