[security] keys: Don't permit request_key() to construct a new keyring
Message-id: <20151020192219.10873.54488.stgit@warthog.procyon.org.uk>
Patchwork-id: 125804
O-Subject: [RHEL7 PATCH 3/3] KEYS: Don't permit request_key() to construct a new keyring
Bugzilla: 1273465
CVE: CVE-2015-7872
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

If request_key() is used to find a keyring, only do the search part - don't
do the construction part if the keyring was not found by the search.  We
don't really want keyrings in the negative instantiated state since the
rejected/negative instantiation error value in the payload is unioned with
keyring metadata.

Now the kernel gives an error:

	request_key("keyring", "#selinux,bdekeyring", "keyring", KEY_SPEC_USER_SESSION_KEYRING) = -1 EPERM (Operation not permitted)

Signed-off-by: David Howells <dhowells@redhat.com>
Original-commit: 911b79cde95c7da0ec02f48105358a36636b7a71
RH-Bugzilla: 1273465
Signed-off-by: Rafael Aquini <aquini@redhat.com>