[powerpc] kvm: book3s_hv: Create debugfs file for each guest's HPT
Message-id: <1445918487-8660-3-git-send-email-dgibson@redhat.com>
Patchwork-id: 125909
O-Subject: [RHEL-7.3 kernel PATCH 2/2] KVM: PPC: Book3S HV: Create debugfs file for each guest's HPT
Bugzilla: 1273692
RH-Acked-by: Laurent Vivier <lvivier@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

This creates a debugfs directory for each HV guest (assuming debugfs
is enabled in the kernel config), and within that directory, a file
by which the contents of the guest's HPT (hashed page table) can be
read.  The directory is named vmnnnn, where nnnn is the PID of the
process that created the guest.  The file is named "htab".  This is
intended to help in debugging problems in the host's management
of guest memory.

The contents of the file consist of a series of lines like this:

  3f48 4000d032bf003505 0000000bd7ff1196 00000003b5c71196

The first field is the index of the entry in the HPT, the second and
third are the HPT entry, so the third entry contains the real page
number that is mapped by the entry if the entry's valid bit is set.
The fourth field is the guest's view of the second doubleword of the
entry, so it contains the guest physical address.  (The format of the
second through fourth fields are described in the Power ISA and also
in arch/powerpc/include/asm/mmu-hash64.h.)

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Alexander Graf <agraf@suse.de>
(cherry picked from commit e23a808b1681d398a983ebc51179efc51c4a1eaf)
Signed-off-by: Rafael Aquini <aquini@redhat.com>

Conflicts:
	arch/powerpc/include/asm/kvm_book3s_64.h

Contextual conflict.

We're porting this patch because it could be useful to track down bugs
1256718 and 1269467.  Given we've had these, it's possible we'll encounter
more bugs where this will be useful.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1273692

Signed-off-by: David Gibson <dgibson@redhat.com>