[mm] ksm: add cond_resched() to the rmap_walks
Message-id: <1446485094-9312-2-git-send-email-aarcange@redhat.com>
Patchwork-id: 126116
O-Subject: [RHEL7.2 PATCH 1/5] mm: add cond_resched() to the rmap walks
Bugzilla: 1274871
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Petr Holasek <pholasek@redhat.com>

The rmap walk must reach every possible mapping of the page, so if a
page is heavily shared (no matter if it's KSM, anon, pagecache) there
will be tons of entries to walk through. All optimizations with
prio_tree, anon_vma chains, interval tree, helps to find the right
virtual mapping faster, but if there are lots of virtual mappings, all
mapping must still be walked through.

The biggest cost is for the IPIs, but regardless of the IPIs, it's
generally safer to keep these cond_resched() in all cases, as even if
we massively reduce the number of IPIs, the number of entries to walk
IPI-less may still be large and no entry can be possibly skipped in
the page migration case.

Acked-by: Hugh Dickins <hughd@google.com>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>