[x86] mm: consolidate VM_FAULT_RETRY handling
Message-id: <1446674187-10981-3-git-send-email-aarcange@redhat.com>
Patchwork-id: 126222
O-Subject: [RHEL7.2 PATCH 2/3] x86: mm: consolidate VM_FAULT_RETRY handling
Bugzilla: 1277226
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

From: Linus Torvalds <torvalds@linux-foundation.org>

The VM_FAULT_RETRY handling was confusing and incorrect for the case of
returning to kernel mode.  We need to handle the exception table fixup
if we return to kernel mode due to a fatal signal - it will basically
look to the kernel user mode access like the access failed due to the VM
going away from udner it.  Which is correct - the process is dying - and
avoids the whole "repeat endless kernel page faults" case.

Handling the VM_FAULT_RETRY early and in just one place also simplifies
the mmap_sem handling, since once we've taken care of VM_FAULT_RETRY we
know that we can just drop the lock.  The remaining accounting and
possible error handling is thread-local and does not need the mmap_sem.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>